+++ AIPScannerConfig/internal/scripts/moduleimport.ps1
@@ -3,15 +3,19 @@
 foreach ($Module in $Moduledependencies) {
 	if (Get-Module -ListAvailable -Name $module) {
 		try {
-			Import-Module -Name $Module -Force -ErrorAction SilentlyContinue	
+			Import-Module -Name $Module -Force -ErrorAction SilentlyContinue
 		}
-		catch {}
+		catch {
+			Write-Error $_.Exception.Message
+		}
 	}
 	else {
 		try {
 			if (-NOT (Install-Module -Name $Module -Scope AllUsers -Force -AllowClobber -ErrorAction Stop -PassThru)) { throw }
 			if (-NOT (Import-Module -Name $Module -Force -ErrorAction SilentlyContinue -PassThru)) { throw }
 		}
-		catch {}
+		catch {
+			Write-Error $_.Exception.Message
+		}
 	}
 }